requirements.txt
requests==2.25.1
flas==1.1.2
pytest==6.2.2